+++ tests/test_ufotest.py
@@ -17,12 +17,14 @@ def response():
     """
     # import requests
     # return requests.get('https://github.com/audreyr/cookiecutter-pypackage')
+    pass
 
 
-def test_content(response):
+def ontent(response):
     """Sample pytest test function with the pytest fixture as an argument."""
     # from bs4 import BeautifulSoup
     # assert 'GitHub' in BeautifulSoup(response.content).title.string
+    pass
 
 
 def command_line_interface():